fix: code review fixes

+++ frame_sorter_test.go
@@ -1419,7 +1419,7 @@ var _ = Describe("frame sorter", func() {
 				BeforeEach(func() {
 					seed := time.Now().UnixNano()
 					fmt.Fprintf(GinkgoWriter, "Seed: %d\n", seed)
-					rand.Seed(seed)
+					rand.Seed(uint64(seed))
 
 					callbacks = nil
 					dataLen = 25
